requirements.txt
Django
djangorestframework
python-decouple
django-filter
djangorestframework-simplejwt
requests
drf-spectacular

